net: shell: Add capture command
Move "capture" command to a separate file.
